fix: clean up markdown review issues
