+++ Sources/Panels/MarkdownPanelView.swift
@@ -302,7 +302,7 @@ private struct MarkdownToolbarButtonStyle: ButtonStyle {
 /// itself when the NSView is created.
 @MainActor
 final class MarkdownWebRendererHandle {
-    fileprivate weak var coordinator: MarkdownWebRenderer.Coordinator?
+    weak var coordinator: MarkdownWebRenderer.Coordinator?
 
     func renderedHTML() async -> String? {
         guard let coordinator else { return nil }
@@ -325,427 +325,5 @@ final class MarkdownWebView: WKWebView {
     override func mouseDown(with event: NSEvent) {
         onPointerDown?()
         super.mouseDown(with: event)
-    }
-}
-
-// MARK: - Web view renderer
-
-private struct MarkdownWebRenderer: NSViewRepresentable {
-    let markdown: String
-    let isDark: Bool
-    let panelId: UUID
-    let workspaceId: UUID
-    let filePath: String
-    let handle: MarkdownWebRendererHandle
-    let onRequestPanelFocus: () -> Void
-
-    func makeCoordinator() -> Coordinator {
-        let c = Coordinator()
-        c.panelId = panelId
-        c.workspaceId = workspaceId
-        c.filePath = filePath
-        handle.coordinator = c
-        return c
-    }
-
-    func makeNSView(context: Context) -> WKWebView {
-        let config = WKWebViewConfiguration()
-        config.suppressesIncrementalRendering = false
-        // Bridge: JS posts to `cmuxLib` to request lazy-loaded libraries
-        // (mermaid / vega-lite). Swift fetches the bundled source from the
-        // app bundle and injects it via evaluateJavaScript.
-        config.userContentController.add(context.coordinator, name: "cmuxLib")
-        let webView = MarkdownWebView(frame: .zero, configuration: config)
-        webView.onPointerDown = onRequestPanelFocus
-        webView.setValue(false, forKey: "drawsBackground")
-        webView.allowsBackForwardNavigationGestures = false
-        webView.allowsLinkPreview = false
-        webView.navigationDelegate = context.coordinator
-        webView.uiDelegate = context.coordinator
-        if #available(macOS 13.3, *) {
-#if DEBUG
-            webView.isInspectable = true
-#else
-            webView.isInspectable = false
-#endif
-        }
-        applyAppearance(to: webView, isDark: isDark)
-
-        context.coordinator.webView = webView
-        context.coordinator.loadShell(isDark: isDark, initialMarkdown: markdown)
-        return webView
-    }
-
-    func updateNSView(_ nsView: WKWebView, context: Context) {
-        // Re-bind the handle in case SwiftUI recreated the representable
-        // wrapper while keeping the same NSView around.
-        handle.coordinator = context.coordinator
-        context.coordinator.panelId = panelId
-        context.coordinator.workspaceId = workspaceId
-        context.coordinator.filePath = filePath
-        (nsView as? MarkdownWebView)?.onPointerDown = onRequestPanelFocus
-        applyAppearance(to: nsView, isDark: isDark)
-        context.coordinator.update(markdown: markdown, isDark: isDark)
-    }
-
-    /// WebKit's `prefers-color-scheme` media query reflects the WKWebView's
-    /// effective NSAppearance. Forcing it here lets us decouple the markdown
-    /// panel from the system appearance and follow the cmux color scheme.
-    private func applyAppearance(to webView: WKWebView, isDark: Bool) {
-        let appearance = NSAppearance(named: isDark ? .darkAqua : .aqua)
-        if webView.appearance !== appearance {
-            webView.appearance = appearance
-        }
-    }
-
-    @MainActor
-    final class Coordinator: NSObject, WKNavigationDelegate, WKUIDelegate, WKScriptMessageHandler {
-        weak var webView: WKWebView?
-        var panelId: UUID = UUID()
-        var workspaceId: UUID = UUID()
-        var filePath: String = ""
-        private var pendingMarkdown: String = ""
-        private var lastMarkdown: String? = nil
-        private var lastIsDark: Bool? = nil
-        private var isLoaded = false
-
-        func loadShell(isDark: Bool, initialMarkdown: String) {
-            pendingMarkdown = initialMarkdown
-            lastIsDark = isDark
-            requestedLibs.removeAll()
-            isLoaded = false
-            let html = MarkdownViewerAssets.shared.shellHTML(isDark: isDark)
-            let baseURL = URL(fileURLWithPath: filePath)
-#if DEBUG
-            NSLog("MarkdownPanel.loadShell filePath=\(filePath) baseURL=\(baseURL.absoluteString) htmlBytes=\(html.utf8.count)")
-#endif
-            webView?.loadHTMLString(html, baseURL: baseURL)
-        }
-
-        func update(markdown: String, isDark: Bool) {
-            let themeChanged = lastIsDark != isDark
-            let contentChanged = lastMarkdown != markdown
-            guard themeChanged || contentChanged else { return }
-
-            pendingMarkdown = markdown
-
-            if themeChanged {
-                lastIsDark = isDark
-                // The WKWebView's NSAppearance change (handled in the
-                // representable's update path) flips `prefers-color-scheme`
-                // automatically. We still nudge the page so highlight.js
-                // swaps stylesheets even if the matchMedia listener is
-                // slow to fire.
-                if isLoaded, let webView {
-                    let js = "window.__cmuxApplyTheme && window.__cmuxApplyTheme();"
-                    webView.evaluateJavaScript(js, completionHandler: nil)
-                }
-            }
-
-            if contentChanged {
-                lastMarkdown = markdown
-                if isLoaded {
-                    pushMarkdown(markdown)
-                }
-            }
-        }
-
-        func renderedHTML() async -> String? {
-            guard let webView, isLoaded else { return nil }
-            // We export an explicit "rendered HTML" getter from JS so callers
-            // get the *content* div only, without the shell <style>/<script>.
-            return await evaluateString("window.__cmuxRenderedHTML && window.__cmuxRenderedHTML()")
-        }
-
-        func renderedText() async -> String? {
-            guard isLoaded else { return nil }
-            return await evaluateString("window.__cmuxRenderedText && window.__cmuxRenderedText()")
-        }
-
-        private func evaluateString(_ script: String) async -> String? {
-            guard let webView else { return nil }
-            do {
-                return try await webView.evaluateJavaScript(script) as? String
-            } catch {
-                return nil
-            }
-        }
-
-        // MARK: Bridge
-
-        private func pushMarkdown(_ markdown: String) {
-            guard let webView else { return }
-#if DEBUG
-            NSLog("MarkdownPanel.pushMarkdown bytes=\(markdown.utf8.count)")
-#endif
-            // Send the raw markdown through a JSON literal so we don't have
-            // to hand-escape backticks/backslashes/quotes for JS.
-            guard let data = try? JSONSerialization.data(withJSONObject: [markdown]),
-                  let arrayLiteral = String(data: data, encoding: .utf8) else { return }
-            let js = """
-            (function(md) {
-              if (window.__cmuxRenderMarkdown) {
-                window.__cmuxRenderMarkdown(md);
-                return;
-              }
-              var el = document.getElementById('content') || document.body;
-              function esc(s) {
-                var div = document.createElement('div');
-                div.textContent = String(s == null ? '' : s);
-                return div.innerHTML;
-              }
-              el.innerHTML = '<pre style=\"color:#f85149;white-space:pre-wrap\">Markdown renderer failed to initialize. Showing raw source.\\n\\n' + esc(md) + '</pre>';
-            })(\(arrayLiteral)[0]);
-            """
-            webView.evaluateJavaScript(js) { _, error in
-#if DEBUG
-                if let error {
-                    NSLog("MarkdownPanel: pushMarkdown evaluateJavaScript failed: \(error)")
-                }
-#endif
-            }
-        }
-
-        // MARK: WKScriptMessageHandler
-
-        func userContentController(
-            _ userContentController: WKUserContentController,
-            didReceive message: WKScriptMessage
-        ) {
-            guard message.name == "cmuxLib",
-                  let body = message.body as? [String: Any] else { return }
-            if let lib = body["lib"] as? String {
-                handleLibRequest(lib)
-                return
-            }
-            if let action = body["action"] as? String {
-#if DEBUG
-                NSLog("MarkdownPanel.bridge action=\(action) body=\(body)")
-#endif
-                switch action {
-                case "resolveMarkdownFile":
-                    guard let requestId = body["requestId"] as? String,
-                          let rawPath = body["path"] as? String else { return }
-                    resolveMarkdownFile(rawPath, requestId: requestId)
-                case "openMarkdownFile":
-                    guard let rawPath = body["path"] as? String else { return }
-                    if let resolved = resolvedMarkdownFilePath(rawPath) {
-                        openMarkdownFile(resolved)
-                    }
-                default:
-                    break
-                }
-            }
-        }
-
-        private var requestedLibs: Set<String> = []
-
-        private func resolveMarkdownFile(_ rawPath: String, requestId: String) {
-            guard let webView else { return }
-            let resolved = resolvedMarkdownFilePath(rawPath)
-#if DEBUG
-            NSLog("MarkdownPanel.resolve raw=\(rawPath) resolved=\(resolved ?? "nil")")
-#endif
-            let payload: [String: Any] = [
-                "requestId": requestId,
-                "exists": resolved != nil,
-                "path": resolved ?? ""
-            ]
-            guard let data = try? JSONSerialization.data(withJSONObject: payload),
-                  let json = String(data: data, encoding: .utf8) else { return }
-            webView.evaluateJavaScript("window.__cmuxMarkdownFileResolved && window.__cmuxMarkdownFileResolved(\(json));", completionHandler: nil)
-        }
-
-        private func resolvedMarkdownFilePath(_ rawPath: String) -> String? {
-            let trimmed = rawPath.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard !trimmed.isEmpty else { return nil }
-            guard MarkdownPanelFileLinkResolver.isMarkdownPathLike(trimmed) else { return nil }
-            return MarkdownPanelFileLinkResolver.resolve(rawPath: trimmed, relativeToMarkdownFile: filePath)
-        }
-
-        private func openMarkdownFile(_ path: String) {
-#if DEBUG
-            NSLog("MarkdownPanel.openMarkdownFile path=\(path)")
-#endif
-            guard let app = AppDelegate.shared,
-                  let location = app.workspaceContainingPanel(
-                      panelId: panelId,
-                      preferredWorkspaceId: workspaceId
-                  ),
-                  let paneId = location.workspace.paneId(forPanelId: panelId) else { return }
-            _ = location.workspace.newMarkdownSurface(
-                inPane: paneId,
-                filePath: path,
-                focus: true
-            )
-        }
-
-        private func handleLibRequest(_ lib: String) {
-            guard let webView else { return }
-            // Load each library at most once per WebView lifetime. If the
-            // shell is reloaded (theme switch), state is reset alongside.
-            if requestedLibs.contains(lib) { return }
-            requestedLibs.insert(lib)
-
-            let assets = MarkdownViewerAssets.shared
-            let sources: [String]
-            switch lib {
-            case "mermaid":
-                sources = [assets.lazyAsset(name: "mermaid.min", ext: "js")]
-            case "vega-lite":
-                // Order matters: vega first, then vega-lite, then vega-embed.
-                sources = [
-                    assets.lazyAsset(name: "vega.min", ext: "js"),
-                    assets.lazyAsset(name: "vega-lite.min", ext: "js"),
-                    assets.lazyAsset(name: "vega-embed.min", ext: "js"),
-                ]
-            default:
-                return
-            }
-
-            // Inject each source as its own script tag via document.head so
-            // that thrown errors get reported through the page console
-            // rather than as a single opaque evaluateJavaScript failure.
-            // Then notify the page that the lib is ready.
-            var injection = ""
-            for src in sources where !src.isEmpty {
-                injection += src
-                injection += "\n;"
-            }
-            // JSON-encode the lib name to safely splice into JS.
-            let libLiteral = (try? JSONSerialization.data(withJSONObject: [lib]))
-                .flatMap { String(data: $0, encoding: .utf8) } ?? "[\"\"]"
-            let suffix = "\nwindow.__cmuxLibLoaded && window.__cmuxLibLoaded(\(libLiteral)[0]);"
-            webView.evaluateJavaScript(injection + suffix) { [weak self] _, error in
-                if let error {
-                    // Allow retry on next render if this attempt failed.
-                    self?.requestedLibs.remove(lib)
-#if DEBUG
-                    NSLog("MarkdownPanel: failed to load \(lib): \(error)")
-#endif
-                }
-            }
-        }
-
-        // MARK: WKNavigationDelegate
-
-        func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-#if DEBUG
-            NSLog("MarkdownPanel.webView.didFinish")
-#endif
-            isLoaded = true
-            // Replay last known markdown after the shell finishes loading.
-            let md = lastMarkdown ?? pendingMarkdown
-            lastMarkdown = md
-            pushMarkdown(md)
-        }
-
-        func webView(
-            _ webView: WKWebView,
-            decidePolicyFor navigationAction: WKNavigationAction,
-            decisionHandler: @escaping (WKNavigationActionPolicy) -> Void
-        ) {
-            // The first load (loadHTMLString) has navigationType = .other —
-            // allow it. Anything the user clicks (links, anchors, ...) we
-            // route through the cmux tab/browser machinery.
-            if navigationAction.navigationType == .linkActivated,
-               let url = navigationAction.request.url {
-#if DEBUG
-                NSLog("MarkdownPanel.nav linkActivated url=\(url.absoluteString)")
-#endif
-                if isInPageFragment(url) {
-                    // Same-document fragment navigation (heading anchors)
-                    // scrolls the panel — keep it native.
-                    decisionHandler(.allow)
-                    return
-                }
-                handleExternalLink(url)
-                decisionHandler(.cancel)
-                return
-            }
-            decisionHandler(.allow)
-        }
-
-        func webView(
-            _ webView: WKWebView,
-            createWebViewWith configuration: WKWebViewConfiguration,
-            for navigationAction: WKNavigationAction,
-            windowFeatures: WKWindowFeatures
-        ) -> WKWebView? {
-            // target=_blank / window.open from inside the rendered markdown.
-            if let url = navigationAction.request.url {
-                handleExternalLink(url)
-            }
-            return nil
-        }
-
-        // MARK: - Link routing
-
-        /// Route a clicked link to a brand-new cmux browser tab in the same
-        /// pane as this markdown panel — mirroring how Browser panels open
-        /// child links via `openLinkInNewTab`. Falls back to the system
-        /// browser only when the in-app browser is disabled or the panel
-        /// can't be located in any workspace.
-        private func handleExternalLink(_ url: URL) {
-#if DEBUG
-            NSLog("MarkdownPanel.handleExternalLink url=\(url.absoluteString)")
-#endif
-            // First preference: links that resolve to local markdown files
-            // open as markdown tabs in cmux, not in the browser.
-            let fileCandidate = url.scheme == "file" ? url.path : url.absoluteString
-            if let markdownPath = resolvedMarkdownFilePath(fileCandidate) {
-                openMarkdownFile(markdownPath)
-                return
-            }
-
-            // Schemes the in-app browser doesn't (and shouldn't) handle:
-            // mailto:, tel:, slack://, vscode://, file:// non-markdown, etc.
-            // Route those to the system handler so the user's default app picks them up.
-            if let scheme = url.scheme?.lowercased(),
-               scheme != "http", scheme != "https" {
-                NSWorkspace.shared.open(url)
-                return
-            }
-
-            guard BrowserAvailabilitySettings.isEnabled() else {
-                NSWorkspace.shared.open(url)
-                return
-            }
-
-            guard let app = AppDelegate.shared,
-                  let location = app.workspaceContainingPanel(
-                      panelId: panelId,
-                      preferredWorkspaceId: workspaceId
-                  ),
-                  let paneId = location.workspace.paneId(forPanelId: panelId) else {
-                // No workspace context — last-resort fallback.
-                NSWorkspace.shared.open(url)
-                return
-            }
-
-            _ = location.workspace.newBrowserSurface(
-                inPane: paneId,
-                url: url,
-                focus: true
-            )
-        }
-
-        private func isInPageFragment(_ url: URL) -> Bool {
-            // Only same-document anchors should stay inside the WebView. With
-            // a file base URL, WebKit resolves `#heading` to
-            // `file:///current.md#heading`; links such as `other.md#heading`
-            // must still route through the markdown-tab opener below.
-            guard url.fragment != nil else { return false }
-            if (url.scheme == nil || url.scheme == "about"), (url.host ?? "").isEmpty {
-                return true
-            }
-            if url.isFileURL {
-                let targetPath = (url.path as NSString).standardizingPath
-                let currentPath = (filePath as NSString).standardizingPath
-                let currentDirectory = ((filePath as NSString).deletingLastPathComponent as NSString).standardizingPath
-                return targetPath == currentPath || targetPath == currentDirectory
-            }
-            return false
-        }
     }
 }
