+++ THIRD_PARTY_LICENSES.md
@@ -182,3 +182,84 @@ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
+
+---
+
+## Markdown Viewer Web Assets
+
+cmux bundles these files under `Resources/markdown-viewer/` so the markdown
+viewer has no runtime CDN dependency.
+
+### marked
+
+- **Version:** 13.0.3
+- **License:** MIT License
+- **Copyright:** Copyright (c) 2011-2024, Christopher Jeffrey
+- **Source:** https://github.com/markedjs/marked
+
+### highlight.js
+
+- **Version:** 11.10.0
+- **License:** BSD 3-Clause License
+- **Copyright:** Copyright (c) 2006-2024 Josh Goebel and other contributors
+- **Source:** https://github.com/highlightjs/highlight.js
+
+### github-markdown-css
+
+- **Version:** bundled CSS from the github-markdown-css package
+- **License:** MIT License
+- **Copyright:** Copyright (c) Sindre Sorhus
+- **Source:** https://github.com/sindresorhus/github-markdown-css
+
+### Mermaid
+
+- **Version:** bundled `mermaid.min.js`
+- **License:** MIT License
+- **Copyright:** Copyright (c) 2014-2024 Knut Sveidqvist and Mermaid contributors
+- **Source:** https://github.com/mermaid-js/mermaid
+
+### Vega
+
+- **Version:** 5.30.0
+- **License:** BSD 3-Clause License
+- **Copyright:** Copyright (c) 2015-2024 University of Washington Interactive Data Lab and contributors
+- **Source:** https://github.com/vega/vega
+
+### Vega-Lite
+
+- **Version:** 5.21.0
+- **License:** BSD 3-Clause License
+- **Copyright:** Copyright (c) 2015-2024 University of Washington Interactive Data Lab and contributors
+- **Source:** https://github.com/vega/vega-lite
+
+### Vega-Embed
+
+- **Version:** 6.26.0
+- **License:** BSD 3-Clause License
+- **Copyright:** Copyright (c) 2015-2024 University of Washington Interactive Data Lab and contributors
+- **Source:** https://github.com/vega/vega-embed
+
+BSD 3-Clause License:
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice, this
+   list of conditions and the following disclaimer.
+2. Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+3. Neither the name of the copyright holder nor the names of its contributors
+   may be used to endorse or promote products derived from this software
+   without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
